refactor(firebase): simplified code by directly calling getAuth in firebase.config.ts
